backend/requirements.txt
Django==5.2.8
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
drf-nested-routers==0.95.0
Pillow==12.0.0
python-dotenv==1.2.1
